Fix parsl resolving proxies twice
Something in Parsl with the HTEX is resolving the input/output proxies
on the client side before the task start execution. Because we have the
evict flag set, the first resolve evicts the data and the task will
fail. Pre-populating the __wrapped__ attribute of the proxy will fix
this while we wait on this feature to make it's way into a ProxyStore
release.